Fixed item iframe source, added config
Fixed a typo in the creation of the url for the embed code which would cause those embed codes to not work. Also cut down on duplication and implemented the embed codes in the admin view, as well as made the appearance of the embed codes in the public view toggle-able.